expenses/pyproject.toml
[project]
name = "expenses-service"
version = "0.1.0"
description = "Expenses microservice built with FastAPI and SQLAlchemy."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "mangum>=0.17.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "mypy>=1.10.0", "ruff>=0.6.0"]

[tool.uvicorn]
app = "app.main:app"
reload = true
